requirements.txt
torch==2.6.0
torchaudio==2.6.0
soundfile
pyqt5
python-vlc
requests
demucs
matplotlib
ml_collections
pillow
rotary_embedding_torch
beartype
einops
librosa
PyYAML